+++ lib/animations/top_left_to_bottom_right_transition.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+import 'package:page_animation_transition/page_animation_interface.dart';
+
+class TopLeftToBottomRightTransition implements PageAnimationInterface {
+  @override
+  Widget animate(BuildContext context, Animation<double> animation,
+      Animation<double> secondaryAnimation, Widget child) {
+    return SlideTransition(
+      position: Tween<Offset>(begin: const Offset(-1, -1), end: Offset.zero)
+          .animate(animation),
+      child: child,
+    );
+  }
+}